add npy encoding
